perf(workflow): reduce doStreamStep step-boundary payload
doStreamStep built the full StepResult inside the step and also returned a
per-chunk snapshot array, serializing both into the durable event log on
every step. The iterator never read the chunk array, and StepResult carries
redundant copies (content, duplicate tool-call lists, reasoningText, the
always-empty *ToolResults arrays).

The step now returns only the minimal raw aggregates needed to rebuild the
StepResult (text, reasoning, responseMetadata, warnings) alongside the
existing finish/toolCalls; the iterator reconstructs the full StepResult via
buildStepResult outside the boundary. The user-facing StepResult shape is
unchanged. runtimeContext/toolsContext/stepNumber no longer cross the
boundary — the iterator applies them during reconstruction.

Ported from vercel/workflow#1932.
